requirements_win.txt
PySide6>=6.6
pyserial>=3.5
numpy>=1.26
matplotlib>=3.8
pyinstaller>=6.3
